convert newField.js into typescript
